clean architecture の適用
