+++ .gitignore
@@ -4,3 +4,4 @@
 /public/edit/*
 /public/index
 /public/index_ul
+/public/db/*
