+++ src/lib.rs
@@ -3,7 +3,10 @@
 extern crate dotenv;
 
 pub mod controllers;
-pub mod models;
+// pub mod models;
+pub mod entities;
 pub mod routes;
+pub mod usecases;
 // pub mod schema;
+pub mod gateways;
 pub mod util;
